fix: move logic page to component
